+++ packages/core/lib/combo/settings/data.ts
@@ -613,6 +613,17 @@ export const SETTINGS = [{
   ],
   default: 'none'
 }, {
+  key: 'moonCrash',
+  name: 'Moon Crash Behavior',
+  category: 'main.events',
+  type: 'enum',
+  description: 'Change the behavior of moon crashing',
+  values: [
+    { value: 'reset', name:  'Reset',  description: 'Moon Crash will restore the last save. No progress will be kept.' },
+    { value: 'cycle', name:  'New Cycle',  description: 'Moon Crash will initiate a new cycle, keeping progress.' },
+  ],
+  default: 'reset'
+}, {
   key: 'startingAge',
   name: 'Starting Age',
   category: 'main.events',